+++ PokerGameApp/PokerGameApp/Model/PokerGame.swift
@@ -15,7 +15,7 @@ class PokerGame {
     private var cardDeck = CardDeck()
     private var sortOfGame: Games
     private var playerCount: PlayerCount // 딜러를 제외한 숫자
-    private(set) var players: [Playable] = []
+    private(set) var players: [Participant] = []
     private(set) var dealer: Dealer = Dealer()
     private var computer: Computer = Computer()
     private var nameArray: [String] = ["HK", "JK", "Crong", "Honux", "Chloe", "Ivy", "Gucci"]
@@ -59,10 +59,6 @@ class PokerGame {
     }
     //TODO: throws를 Result<>로 변환을 해야할 거같은데.. 인터넷에 나와있는 예제들은 비동기 함수에서 적용하는 법을 알려줘서 지금과 같은 동기 함수에서는 어떻게 적용해야할지 모르겠다. https://onelife2live.tistory.com/1 잘 정리되어있음.
     func giveCard2Player() {
-        guard round != 7 else {
-            print("카드를 모두 배분했습니다.")
-            return
-        }
         round += 1
         for player in players {
             if let selectedCard = try? cardDeck.removeOne().get() {
@@ -73,4 +69,5 @@ class PokerGame {
             }
         }
     }
+    
 }
